The lifecycle of rating system is almost done.

+++ Tigra/Database/Tables/RequirementRatings.sql
@@ -4,9 +4,9 @@ BEGIN
 	CREATE TABLE [Tigra].[RequirementRatings]
 	(
 		[RevisionID] BIGINT NOT NULL,
-		[RatingA] TINYINT NOT NULL CHECK([RatingA] BETWEEN 0 AND 5),
-		[RatingB] TINYINT NOT NULL CHECK([RatingB] BETWEEN 0 AND 5),
-		[RatingC] TINYINT NOT NULL CHECK([RatingC] BETWEEN 0 AND 5),
+		[RatingA] DECIMAL(3,1) NOT NULL CHECK([RatingA] BETWEEN 0 AND 5),
+		[RatingB] DECIMAL(3,1) NOT NULL CHECK([RatingB] BETWEEN 0 AND 5),
+		[RatingC] DECIMAL(3,1) NOT NULL CHECK([RatingC] BETWEEN 0 AND 5),
 		[FinalRating] AS (([RatingA] + [RatingB] + [RatingC]) / 3.0),
 		[Approved] BIT NOT NULL,
 		CONSTRAINT [pk_Tigra_RequirementRatings_Key] PRIMARY KEY ([RevisionID]),
